update alias target to paddle.device.cpu.get_rng_state
